receipt db:price change integer to float

+++ db/migrate/20171216082402_create_receipts.rb
@@ -4,7 +4,7 @@ class CreateReceipts < ActiveRecord::Migration[5.0]
       t.string :customer_name
       t.string :payment_method
     	t.string :des
-      t.integer :price, null: false, default: 0
+      t.float :price, null: false, default: 0
       t.date :date
 
       t.timestamps
